Marshall-Rosenberg-Quotes/facebook/requirements.txt
facebook-sdk==2.0.0
pyyaml